Add a Jump Register (#23)
After some thought, it feels better to have a dedicated Jump register built into the program counter. This helps deal with the design having 8-bit registers but using 16-bit addressing, which had lead to needing the instruction to copy the PC to be able to write to *two* registers. It was the only instruction which needed to do so, and complicated the register file. Having a jump register avoids this, since it is safe to copy the low and high bytes with separate instructions. I believe that this compensates for the extra instructions now in the Program Counter.

Closes #19