Expand assembler for new test (#29)
Create a new program to count by five. This uses the DALU and also stores its result in main memory. Consequently, the assembler has been expanded to cope with the instructions this requires. Change the E2E tests to load the sample programs, rather than having code embedded in the test file.