Fix Tester board (#49)
The 'input' level shifter on the Tester board had the wrong set of pins connected to ground (copy paste error from the 'output' level shifters). This meant that the IC could effectively short circuit and become very hot. Switch over the ground and floating pins, to reflect how the inputs and outputs are for this IC.

Closes #47